# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UPS e-Recruit API",
    "version": "1.0.0",
    "description": "Versioned same-origin API for applicant and authorised recruitment operations. OCR is assistive evidence only."
  },
  "servers": [{ "url": "/api" }],
  "security": [{ "sanctum": [] }],
  "paths": {
...
    "/v1/final-selections": { "post": { "operationId": "approveFinalSelection", "requestBody": { "$ref": "#/components/requestBodies/JsonBody" }, "responses": { "201": { "description": "Council-approved selected and Fit candidate" }, "409": { "$ref": "#/components/responses/Conflict" }, "422": { "$ref": "#/components/responses/ValidationError" } } } },
...
  },
  "components": {
    "securitySchemes": { "sanctum": { "type": "http", "scheme": "bearer" } },
...
    "requestBodies": { "JsonBody": { "required": true, "content": { "application/json": { "schema": { "type": "object" } } } } },
    "responses": {
...
      "Conflict": { "description": "Version, offline, or workflow gate conflict" },
      "ValidationError": { "description": "Input validation failed" },
...
    }
  }
}